requirements.txt
python-dotenv
discord.py
PyNaCl
psycopg2
